Added NSLocationAlwaysUsageDescription which is needed for iOS 8.
In iOS 8, you will need to add a new key (NSLocationAlwaysUsageDescription or NSLocationWhenInUseUsageDescription) in your PList to get the location update. NSLocationAlwaysUsageDescription is used when you will need the location update all the time even when the app is in the background while NSLocationWhenInUseUsageDescription is needed only if you need the location update when the app is in the foreground.